requirements.txt
flask
flask-cors
mysql-connector-python
pandas
scapy
psutil
netifaces